chipiq/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
deepface==0.0.93
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.3
statsmodels==0.14.4
prophet==1.1.6
cmdstanpy==1.2.5
scikit-learn==1.5.2
tensorflow==2.17.0
python-multipart==0.0.12
Pillow==10.4.0
tf-keras==2.17.0
kafka-python==2.0.2
pyspark==3.5.2
psycopg[binary]==3.2.1
